Ajuste nos Controller do Comparativo, Cadastro e Comuns  - Realizada revisão dos controllers da sessao de cadastro, com ajustes nas funcionalidades, alem de comentarios para melhor compreensao das rotinas.  - Reorganizacao dos controllers de Registro, Termo, Sobre e Usuario, isolando os mesmos dos controllers comuns, e ajustando as rotas quanto a importacao dos presentes controllers.  - Criada classes e Metodos estaticos para busca de dados no BD, utilizando Cache, de dados gerais do Comparativo.  - Criada classes e Metodos estaticos para busca de dados no BD, utilizando Cache, a nivel de Municipio e Escola.  - Alterado controller de Comparativo a nivel de Municipio e Escola, para fins de utilizar as classes e Metodos Estaticos criados.  - Alterado controller de Manutencao de Comparativo a nivel de Municipio e Escola, visando utilizar as classes e Metodos Estaticos criados.
