+++ app/Http/Controllers/HomeController.php
@@ -2,23 +2,18 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\DestaqueModel;
-use App\Models\Escola;
 use App\Models\Legenda;
-use App\Models\Municipio;
 use App\Models\Previlegio;
-use App\Models\Prova;
-use App\Models\Prova_gabarito;
-use App\Models\Questao;
 use App\Models\Solicitacao;
 use App\Models\Sugestao;
-use App\Models\Turma;
-use App\Models\User;
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\DB;
 
 class HomeController extends Controller
 {
+    private $objPrevilegio;
+    private $objSolicitacao;
+    private $objLegenda;
+    private $objSugestao;
+
     /**
      * Create a new controller instance.
      * Método construtor que realiza a inicilização dos objetos e classes que serão utilizados na conexão
@@ -28,7 +23,6 @@ class HomeController extends Controller
     public function __construct()
     {
         $this->middleware('auth');
-        $this->objUser = new User();
         $this->objPrevilegio = new Previlegio();
         $this->objSolicitacao = new Solicitacao();
         $this->objLegenda = new Legenda();
@@ -48,11 +42,16 @@ class HomeController extends Controller
 
         //Caso seja administrados tem acesso a todas as solicitações em aberto
         if (auth()->user()->perfil == 'Administrador') {
+
+            //Busca listagem de Solicitações gerais
             $solRegistro = $this->objSolicitacao->where(['aberto' => '1'])->where(['id_tipo_solicitacao' => 1])->get();
             $solAltCadastral = $this->objSolicitacao->where(['aberto' => '1'])->where(['id_tipo_solicitacao' => 2])->get();
             $solAddTurma = $this->objSolicitacao->where(['aberto' => '1'])->where(['id_tipo_solicitacao' => 3])->get();
+
             //Caso seja gestor, tem acesso a todas as solicitações do município que esta vinculado
         } else if (isset($previlegio[0]->funcaos_id) && $previlegio[0]->funcaos_id == 6) {
+
+            //Busca listagem de Solicitações pelo Município definido no Previlégio do Usuário logado
             $solRegistro = $this->objSolicitacao->where(['aberto' => '1'])->where(['id_tipo_solicitacao' => 1, 'id_municipio' => $previlegio[0]->municipios_id])->get();
             $solAltCadastral = $this->objSolicitacao->where(['aberto' => '1'])->where(['id_tipo_solicitacao' => 2, 'id_municipio' => $previlegio[0]->municipios_id])->get();
             $solAddTurma = $this->objSolicitacao->where(['aberto' => '1'])->where(['id_tipo_solicitacao' => 3, 'id_municipio' => $previlegio[0]->municipios_id])->get();
@@ -61,18 +60,23 @@ class HomeController extends Controller
         //Caso não tenha previlégio adicionados ao usuário, acessa apenas a página inicial de boas vinda
         if (!isset($previlegio[0]->funcaos_id)) {
             return view('bem-vindo', compact('legendas', 'sugestoes'));
+
             //Caso seja gestor acessa apenas a página de boas vindos, com as solicitações disponíveis
         } else if ((isset($previlegio[0]->funcaos_id) && $previlegio[0]->funcaos_id == 6)) {
             return view('bem-vindo', compact('solRegistro', 'solAltCadastral', 'solAddTurma', 'legendas', 'sugestoes'));
+
             //Caso seja Administrador, acessa todo conteúdo do site, iniciando pelça página do secretario
         } else if (auth()->user()->perfil == 'Administrador') {
             return redirect()->route('secretario.index');
+
             //Caso seja secretario ou pesquisador do munícipio unijuí, inicia pela página de secretario
         } else if ($previlegio[0]->funcaos_id == 8 || (($previlegio[0]->funcaos_id == 13 || $previlegio[0]->funcaos_id == 14) && $previlegio[0]->municipios_id == 5)) {
             return redirect()->route('secretario.index');
+
             //Caso seja diretor, inicia pela página do diretor
         } else if ($previlegio[0]->funcaos_id == 5) {
             return redirect()->route('diretor.index');
+
             //Caso seja professor, inicia pela página de professor
         } else if ($previlegio[0]->funcaos_id == 7) {
             return redirect()->route('professor.index');
