+++ app/Http/Controllers/cadastros/aluno/AlunoFilterController.php
@@ -12,14 +12,13 @@ use App\Models\AnoSame;
 
 class AlunoFilterController extends Controller
 {
+    private $objAnoSame;
+
      /**
      * Método construtor que inicializa as classes a serem utilizadas para ações de comunicação com o banco de dados
      */
     public function __construct()
     {
-        $this->objTurma = new Turma();
-        $this->objEscola = new Escola();
-        $this->objAluno = new Aluno();
         $this->objAnoSame = new AnoSame();
     }
     /**
@@ -49,9 +48,7 @@ class AlunoFilterController extends Controller
                         ->join('turmas', ['alunos.turmas_id' => 'turmas.id', 'alunos.SAME' => 'turmas.SAME']) 
                         ->select('alunos.*', 'municipios.nome as nome_municipio','escolas.nome as nome_escola','turmas.TURMA as nome_turma')
                         ->orderBy('updated_at', 'desc')->paginate(7);
-        //$escolas = $this->objEscola->all();
         $anossame = $this->objAnoSame->orderBy('descricao', 'asc')->get();
-        //$turmas = $this->objTurma->all();
 
         return view('cadastro/aluno/list_aluno', compact('alunos','anossame'));
     }
